Resizable datagrid: Refactored ExecuteAfterRender into its own component. Refactored execution to use dequeue and not allocate a new array in memory.
