+++ Source/Blazorise/Base/BaseAfterRenderComponent.cs
@@ -1,0 +1,71 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Components;
+
+namespace Blazorise.Base
+{
+    public abstract class BaseAfterRenderComponent : ComponentBase, IDisposable
+    {
+
+        /// <summary>
+        /// Indicates if component has been rendered in the browser.
+        /// </summary>
+        protected bool Rendered { get; private set; }
+
+        /// <summary>
+        /// A stack of functions to execute after the rendering.
+        /// </summary>
+        protected Queue<Func<Task>> executeAfterRenderQueue;
+
+        /// <summary>
+        /// Pushes an action to the stack to be executed after the rendering is done.
+        /// </summary>
+        /// <param name="action"></param>
+        protected void ExecuteAfterRender( Func<Task> action )
+        {
+            if ( executeAfterRenderQueue == null )
+                executeAfterRenderQueue = new Queue<Func<Task>>();
+
+            executeAfterRenderQueue.Enqueue( action );
+        }
+
+        protected override async Task OnAfterRenderAsync( bool firstRender )
+        {
+            Rendered = true;
+            if ( executeAfterRenderQueue?.Count > 0 )
+            {
+                while ( executeAfterRenderQueue.Count > 0 )
+                {
+                    await executeAfterRenderQueue.Dequeue()();
+                }
+            }
+
+            await base.OnAfterRenderAsync( firstRender );
+        }
+
+        private bool disposedValue;
+
+        protected virtual void Dispose( bool disposing )
+        {
+            if ( !disposedValue )
+            {
+                if ( disposing )
+                {
+                    executeAfterRenderQueue.Clear();
+                    executeAfterRenderQueue = null;
+                }
+
+                disposedValue = true;
+            }
+        }
+
+        public void Dispose()
+        {
+            Dispose( disposing: true );
+            GC.SuppressFinalize( this );
+        }
+    }
+}
